Add ability to use clock::parse_date_time to load datetimes
Base R does not handle ambiguous or nonexistent datetimes nicely - ambiguous datetimes (e.g. when BST transitions to GMT) may end up as either BST or GMT (it seems to be platform dependent and not to be even deterministic). clock makes all of this explicit.

The code works by rewriting the cols() supplied to map col_datetime() to col_character(), so that the raw data are imported as characters. We then parse the character data into POSIXct using clock::parse_date_time