Rewrite of `summarise_during()`, `all_during()` etc
Major speedups are enabled, especially on database backends by constructing a spec for pivotting wider. This avoids
- the need to use `split_by_type()`, which is very slow on huge datasets
- the need for `dbplyr::pivot_wider()` to compute the new column names by itself, which is also slow

Joining in `all_during()` is now directly handled as an inequality join in `dplyr`, which should be faster than the old implementation.

Default join type for `all_during()` is now an inner join to avoid surprising NAs. The `datetime` column is no longer renamed in `all_during()`.

Also adds many new tests of the`pivo non-lazy-tbl versions of the join functions are added. Tests of the dbplyr versions are needed, particularly given `pivot_value_wider_spec()` uses different methods for database backends

Original implementations of all of these functions are retained (unexported) as `v1_*()`. Need to add tests to compare the outputs of the new and old versions